Moving pollers out of avsExecutionManager.
